Add configuration file to load assets
Meson will automatically generate a file in `headers/config.hpp`,
containing a `define PACKAGE_DATADIR "/path"`. This way, the path
where the assets will be stored can be determined at compile-time.

It is simple to create a config file if Meson is not used. Meson will
not overwrite the file if it exists.